requirements.txt
langchain>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.1.7
openai>=1.30.0
python-dotenv>=1.0.0
sqlalchemy>=2.0
pandas
